Add mamba_18M baseline, clean up training loop and configs
- Add MAMBA_18M config (d=512, L=10, 17.5M params)
- Fix DataLoader uint16 -> int for embedding compatibility
- Clean up logging: f-string format, tokens as M, no lr in output
- Fix double eval/save on last step
- Gitignore checkpoints (too large for GitHub)

+++ data/loader.py
@@ -198,9 +198,9 @@ class DataLoader:
     def batch(self) -> tuple[torch.Tensor, torch.Tensor]:
         ix = torch.randint(len(self.data) - self.T, (self.B,))
         x = torch.stack(
-            [torch.from_numpy(self.data[i : i + self.T].copy()) for i in ix]
+            [torch.from_numpy(self.data[i : i + self.T].copy().astype(int)) for i in ix]
         )
         y = torch.stack(
-            [torch.from_numpy(self.data[i + 1 : i + 1 + self.T].copy()) for i in ix]
+            [torch.from_numpy(self.data[i + 1 : i + 1 + self.T].copy().astype(int)) for i in ix]
         )
         return x, y
